+++ src/FormFactory.php
@@ -132,12 +132,21 @@ final class FormFactory implements FormFactoryInterface
             }
         }
 
+        // Get the submit button
+        $submitName = null;
+        foreach ($this->getSubmitStateNodeList() as $name) {
+            if (isset($data[$name])) {
+                $submitName = $name;
+            }
+        }
+
         // Return validation result
         return new ValidationResult(
             $inputFilter->getRawValues(),
             $inputFilter->getValues(),
             $messages,
-            $method
+            $method,
+            $submitName
         );
     }
 
@@ -192,13 +201,33 @@ final class FormFactory implements FormFactoryInterface
                 $name = $node->getAttribute('data-input-name');
             }
 
-            if (!$name) {
+            if (!$name || $node->getAttribute('type') == 'submit') {
                 // At least a name is needed to submit a value.
                 // Silently continue, might be a submit button.
                 continue;
             }
 
             yield $name => $node;
+        }
+    }
+
+    /**
+     * Get names of available named submit elements
+     */
+    private function getSubmitStateNodeList()
+    {
+        $xpath = new DOMXPath($this->document);
+        $nodeList = $xpath->query('//input[@type="submit"] | //button[@type="submit"]');
+
+        /** @var DOMElement $node */
+        foreach ($nodeList as $node) {
+            $name = $node->getAttribute('name');
+            if (!$name) {
+                // At least a name is needed to submit a value.
+                continue;
+            }
+
+            yield $name;
         }
     }