+++ src/ValidationResultInterface.php
@@ -11,8 +11,15 @@ interface ValidationResultInterface
      * @param array       $validatedData
      * @param array       $errors
      * @param null|string $method
+     * @param null|string $submitName
      */
-    public function __construct(array $rawInputData, array $validatedData, array $errors, $method = null);
+    public function __construct(
+        array $rawInputData,
+        array $validatedData,
+        array $errors,
+        $method = null,
+        $submitName = null
+    );
 
     /**
      * Check if the validation was successful
@@ -23,6 +30,18 @@ interface ValidationResultInterface
      * @return bool
      */
     public function isValid();
+
+    /**
+     * Checks if submit button is clicked or return its name
+     *
+     * If a specific name is given it checks if a submit button is clicked and returns a boolean.
+     * In case no name is given, it returns the name of the clicked button or null if no named was clicked.
+     *
+     * @param null|string $name
+     *
+     * @return null|boolean|string
+     */
+    public function isClicked($name = null);
 
     /**
      * Get validation messages