+++ CHANGELOG.md
@@ -8,6 +8,11 @@ All notable changes to this project will be documented in this file, in reverse 
 
 - [#36](https://github.com/xtreamwayz/html-form-validator/pull/36) adds the `FormFactory->validateRequest()` to handle
   PSR-7 requests and reduce boilerplate code needed to build, validate and render forms.
+- [#38](https://github.com/xtreamwayz/html-form-validator/pull/38) adds submit button detection.
+  ```php
+  $validationResult->isClicked('confirm'); // returns boolean;
+  $validationResult->isClicked(); // returns name of clicked button or null;
+  ```
 
 ### Deprecated
 
@@ -19,7 +24,8 @@ Nothing.
 
 ### Fixed
 
-Nothing.
+- [#38](https://github.com/xtreamwayz/html-form-validator/pull/38) fixes named submit buttons being detected as a
+  validating input value.
 
 ## 0.3.0 - 2016-02-22
 
@@ -35,7 +41,7 @@ Nothing.
     - min, max and step attributes for Datetime input types
     - the multiple attribute on select and email elements
     - file element validation options
-     
+
 ### Deprecated
 
 - [#27](https://github.com/xtreamwayz/html-form-validator/pull/26) deprecates: